+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix possible Cross-site Scripting (XSS) in Rest/GraphQL viewer
 - Add src setting for slect and tag fields to load options from internal sources
 - Sanitize space name during installation
+- Multiselect & delete of assets
 
 ## 2.6.3 (2023-08-15)